+++ src/main/java/com/georgster/control/manager/ChatCompletionManager.java
@@ -10,6 +10,7 @@ import com.georgster.control.util.ClientContext;
 import com.georgster.database.ProfileType;
 import com.georgster.gpt.MemberChatCompletions;
 import com.theokanning.openai.completion.chat.ChatCompletionRequest;
+import com.theokanning.openai.completion.chat.ChatCompletionResult;
 import com.theokanning.openai.completion.chat.ChatMessage;
 import com.theokanning.openai.service.OpenAiService;
 
@@ -49,13 +50,6 @@ public class ChatCompletionManager extends SoapManager<MemberChatCompletions> {
     }
 
     /**
-     * Loads all previous chat completions into this manager.
-     */
-    public void load() {
-        dbService.getAllObjects().forEach(observees::add);
-    }
-
-    /**
      * Creates a new ChatCompletion for the provided {@code Member} based on the given
      * prompt and returns the first response.
      * The response will be based on the previous ten chat completions for that member
@@ -66,27 +60,9 @@ public class ChatCompletionManager extends SoapManager<MemberChatCompletions> {
      * @return The response from the AI.
      */
     public String createCompletion(String prompt, Member member) {
-        String id = member.getId().asString();
+        MemberChatCompletions completions = get(member.getId().asString());
 
-        add(new MemberChatCompletions(id));
-
-        MemberChatCompletions completions = get(id);
-
-        List<ChatMessage> messages = new ArrayList<>();
-        messages.add(new ChatMessage("system", "You are a Discord bot called SOAP Bot."));
-
-        completions.getTokens().forEach(token -> 
-            token.forEach((k,v) -> {
-                messages.add(new ChatMessage("user", k));
-                messages.add(new ChatMessage("assistant", v));
-            })
-        );
-
-        messages.add(new ChatMessage("user", prompt));
-
-        ChatCompletionRequest request = ChatCompletionRequest.builder().messages(messages).model("gpt-3.5-turbo").build();
-
-        String response = aiService.createChatCompletion(request).getChoices().get(0).getMessage().getContent();
+        String response = createCompletionRequest(prompt, member).getChoices().get(0).getMessage().getContent();
         completions.addCompletion(prompt, response);
 
         update(completions);
@@ -108,18 +84,34 @@ public class ChatCompletionManager extends SoapManager<MemberChatCompletions> {
      * @return All responses from the AI.
      */
     public List<String> createCompletionGetAll(String prompt, Member member) {
+        MemberChatCompletions completions = get(member.getId().asString());
+
+        List<String> responses = new ArrayList<>();
+
+        createCompletionRequest(prompt, member).getChoices().forEach(choice -> responses.add(choice.getMessage().getContent()));
+        completions.addCompletion(prompt, responses.get(0));
+
+        update(completions);
+
+        return responses;
+    }
+
+    /**
+     * Creates a ChatCompletionRequest to OpenAI using the {@code gpt-3.5-turbo} model.
+     * 
+     * @param prompt The prompt from the user.
+     * @param member The member who prompted the completion request.
+     * @return The result of the request.
+     */
+    private ChatCompletionResult createCompletionRequest(String prompt, Member member) {
         String id = member.getId().asString();
 
         add(new MemberChatCompletions(id));
 
-        MemberChatCompletions completions = get(id);
-
         List<ChatMessage> messages = new ArrayList<>();
-        messages.add(new ChatMessage("system", "You are a Discord bot called SOAP Bot who was created by georgster." +
-                                                    "You have many features, but if anyone is confused on how to use the bot, tell them to use !help" +
-                                                    "or use slash commands (with '/') to bring up a menu of commands."));
+        messages.add(new ChatMessage("system", "You are a Discord bot called SOAP Bot."));
 
-        completions.getTokens().forEach(token -> 
+        get(id).getTokens().forEach(token -> 
             token.forEach((k,v) -> {
                 messages.add(new ChatMessage("user", k));
                 messages.add(new ChatMessage("assistant", v));
@@ -130,13 +122,6 @@ public class ChatCompletionManager extends SoapManager<MemberChatCompletions> {
 
         ChatCompletionRequest request = ChatCompletionRequest.builder().messages(messages).model("gpt-3.5-turbo").build();
 
-        List<String> responses = new ArrayList<>();
-
-        aiService.createChatCompletion(request).getChoices().forEach(choice -> responses.add(choice.getMessage().getContent()));
-        completions.addCompletion(prompt, responses.get(0));
-
-        update(completions);
-
-        return responses;
+       return aiService.createChatCompletion(request);
     }
 }